apply typed enum

+++ bits64.go
@@ -1,3 +1,14 @@
+// Copyright 2015 SeukWon Kang (kasworld@gmail.com)
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//    http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 //goroutine safe uint64 bit operation
 package bits64
 
@@ -12,7 +23,7 @@ func (t Bits64) String() string {
 	return fmt.Sprintf("0b%0b", t)
 }
 
-func (bt *Bits64) SetBit(n uint) {
+func (bt *Bits64) SetBit(n uint8) {
 	var success bool
 	for !success {
 		oldval := atomic.LoadUint64((*uint64)(bt))
@@ -21,7 +32,7 @@ func (bt *Bits64) SetBit(n uint) {
 	}
 }
 
-func (bt *Bits64) ClearBit(n uint) {
+func (bt *Bits64) ClearBit(n uint8) {
 	var success bool
 	for !success {
 		oldval := atomic.LoadUint64((*uint64)(bt))
@@ -30,7 +41,7 @@ func (bt *Bits64) ClearBit(n uint) {
 	}
 }
 
-func (bt *Bits64) NegBit(n uint) {
+func (bt *Bits64) NegBit(n uint8) {
 	var success bool
 	for !success {
 		oldval := atomic.LoadUint64((*uint64)(bt))
@@ -39,7 +50,7 @@ func (bt *Bits64) NegBit(n uint) {
 	}
 }
 
-func (bt *Bits64) GetBit(n uint) bool {
+func (bt *Bits64) GetBit(n uint8) bool {
 	val := atomic.LoadUint64((*uint64)(bt)) & (1 << n)
 	return val != 0
 }
